#8 - migrate datalake module to global db subpackage
