+++ main-package/db/utils.py
@@ -1,6 +1,6 @@
 from typing import List
 from sqlalchemy.orm import sessionmaker
-from datalake.tables import *
+
 class TableInterface:
     pass
 
@@ -23,10 +23,4 @@ def add_rows(engine, table:TableInterface ,rows:List[RowInterface]):
         session.bulk_save_objects(user_list)
         session.commit()
 
-def add_all_homepage(clean_table_data, engine):
-    added_data = (HomePageTableFactory(**i) for i in clean_table_data)
-    add_rows(engine, Table, added_data)
 
-def add_all_additional_info(clean_src_data, engine):
-    added_data = (AdditionalInfoFactory(*i) for i in clean_src_data)
-    add_rows(engine, AdditionalInfo, added_data)
